feat: implement content area loader and refactor layout components
